feat(route): add route handler for deleting martial art
