+++ server.js
@@ -6,11 +6,9 @@ const script = fs.readFileSync("./public/client.js", "utf-8")
 const styles = fs.readFileSync("./public/styles.css", "utf-8")
 const UI = fs.readFileSync("./public/marts.html", "utf-8")
 
-const { createMartialArt, getAllMartialArts, deleteMartialArt, updateDescription } = require("./database.js")
-const { handleCreateRequest, handleReadRequest } = require("./routes/martial-arts.route.js")
+const { handleCreateRequest, handleReadRequest, handleUpdateRequest, handleDeleteRequest } = require("./routes/martial-arts.route.js")
 const server = http.createServer(async (req, res) => {
 
-    // READ
     if (req.url === "/martial-arts" && req.method === "GET") {
         res.writeHead(200, { "Content-Type": "text/html" })
         res.end(UI)
@@ -21,35 +19,11 @@ const server = http.createServer(async (req, res) => {
     }
 
     else if (req.url === "/delete" && req.method === "DELETE") {
-        let body = ""
-        req.on("data", chunk => {
-            body += chunk.toString()
-        })
-        req.on('end', async () => {
-            const bodyData = JSON.parse(body)
-            const { id } = bodyData
-            const success = await deleteMartialArt(id)
-            if (success) {
-                res.writeHead(200, { "Content-Type": "application/json" })
-                res.end(id)
-            }
-        });
+        handleDeleteRequest(req, res)
     }
 
     else if (req.url === "/update" && req.method === "PATCH") {
-        let body = ""
-        req.on("data", chunk => {
-            body += chunk.toString()
-        })
-        req.on('end', async () => {
-            const bodyData = JSON.parse(body)
-            const { id, newDesc } = bodyData
-            const success = await updateDescription(newDesc, id)
-            if (success) {
-                res.writeHead(200, { "Content-Type": "application/json" })
-                res.end()
-            }
-        });
+        handleUpdateRequest(req, res)
     }
 
     // Route for getting client.js file
